requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
pyyaml==6.0.1
openai==1.12.0
anthropic==0.18.0
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.2
apscheduler==3.10.4
watchdog==3.0.0
requests==2.31.0
